Docstrings

WhittleLikelihoodInference.AdditiveTimeSeriesModel

WhittleLikelihoodInference.BartlettPeriodogram

WhittleLikelihoodInference.DebiasedWhittleData

WhittleLikelihoodInference.Periodogram

WhittleLikelihoodInference.TimeSeriesModel

WhittleLikelihoodInference.TimeSeriesModelStorage

WhittleLikelihoodInference.UnknownAcvTimeSeriesModel

WhittleLikelihoodInference.WhittleData

Base.:+

Base.:+

Base.ndims

WhittleLikelihoodInference.EI

WhittleLikelihoodInference.EI!

WhittleLikelihoodInference.acv

WhittleLikelihoodInference.acv

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.acv!

WhittleLikelihoodInference.add_asdf!

WhittleLikelihoodInference.add_sdf!

WhittleLikelihoodInference.asdf

WhittleLikelihoodInference.asdf!

WhittleLikelihoodInference.asdf!

WhittleLikelihoodInference.checkparameterlength

WhittleLikelihoodInference.checkposreal

WhittleLikelihoodInference.coherance

WhittleLikelihoodInference.coherancy

WhittleLikelihoodInference.debiasedwhittle!

WhittleLikelihoodInference.debiasedwhittle_Ehess!

WhittleLikelihoodInference.debiasedwhittle_FG!

WhittleLikelihoodInference.debiasedwhittle_FGH!

WhittleLikelihoodInference.debiasedwhittle_Fisher!

WhittleLikelihoodInference.extract_EI

WhittleLikelihoodInference.extract_acv

WhittleLikelihoodInference.extract_array

WhittleLikelihoodInference.fit

WhittleLikelihoodInference.getallderiv

WhittleLikelihoodInference.grad_EI!

WhittleLikelihoodInference.grad_acv

WhittleLikelihoodInference.grad_acv!

WhittleLikelihoodInference.grad_add_asdf!

WhittleLikelihoodInference.grad_add_sdf!

WhittleLikelihoodInference.grad_asdf!

WhittleLikelihoodInference.grad_asdf!

WhittleLikelihoodInference.grad_sdf

WhittleLikelihoodInference.groupdelay

WhittleLikelihoodInference.hess_EI!

WhittleLikelihoodInference.hess_acv!

WhittleLikelihoodInference.hess_add_asdf!

WhittleLikelihoodInference.hess_add_sdf!

WhittleLikelihoodInference.hess_asdf!

WhittleLikelihoodInference.hess_asdf!

WhittleLikelihoodInference.hess_sdf

WhittleLikelihoodInference.indexLT

WhittleLikelihoodInference.minbins

WhittleLikelihoodInference.nalias

WhittleLikelihoodInference.nlowertriangle_dimension

WhittleLikelihoodInference.nlowertriangle_parameter

WhittleLikelihoodInference.parameternames

WhittleLikelihoodInference.plotacv

WhittleLikelihoodInference.plotasdf

WhittleLikelihoodInference.plotcoh

WhittleLikelihoodInference.plotei

WhittleLikelihoodInference.plotsdf

WhittleLikelihoodInference.sdf

WhittleLikelihoodInference.whittle!

WhittleLikelihoodInference.whittle_FG!

WhittleLikelihoodInference.whittle_FGH!

WhittleLikelihoodInference.AdditiveTimeSeriesModel — Type

AdditiveTimeSeriesModel(model1, model2)

Constructs an additive model from two timeseries models.

WhittleLikelihoodInference.BartlettPeriodogram — Type

BartlettPeriodogram(timeseries, Δ, segmentlength)

Compute the Bartlett periodogram for the provided timeseries with sampling rate Δ.

Arguments

timeseries: A Vector if univariate and an n by d Matrix if multivariate, where n is the number of observations and d is the dimension of the timeseries.

Δ: A positive real number.

segmentlength: the length of series used in each segment.

Computes an estimate of the spectral density function using Bartlett's method. Using the same normalisation as Periodogram.

External links

Bartlett's method on Wikipedia

WhittleLikelihoodInference.DebiasedWhittleData — Type

DebiasedWhittleData(model::Type{<:TimeSeriesModel{D}}, timeseries, Δ; lowerΩcutoff, upperΩcutoff, taper)

Create storage for timeseries data in a format which is useful for debiased Whittle methods.

WhittleLikelihoodInference.Periodogram — Type

Periodogram(timeseries, Δ)

Compute the periodogram for the provided timeseries with sampling rate Δ.

Arguments

timeseries: A Vector if univariate and an n by d Matrix if multivariate, where n is the number of observations and d is the dimension of the timeseries.

Δ: A positive real number.

The periodogram is defined as

\[\boldsymbol I(ω)&=\boldsymbol J(ω) \boldsymbol J(ω)^H \quad \text{where} \quad \boldsymbol J(ω) = \sqrt{\frac{Δ}{2π n}}\sum_{t=0}^{n-1} \boldsymbol{P}_{tΔ}e^{-itΔ ω}\]

Note the periodogram is in terms of angular frequency here, and uses the normalisation $Δ/2π$. The choice of normalisation is essentially arbitrary; however, this matches our definition for the spectral density function.

WhittleLikelihoodInference.TimeSeriesModel — Type

Time series model with a parameter for the dimension of the timeseries model.

WhittleLikelihoodInference.TimeSeriesModelStorage — Type

TimeSeriesModelStorage

Preallocated storage for model manipulation including preplanned ffts.

WhittleLikelihoodInference.UnknownAcvTimeSeriesModel — Type

Sub type to specify the need for additional routines approximating the acv from the sdf. Such routines require additional storage and have different default functions in some cases.

WhittleLikelihoodInference.WhittleData — Type

WhittleData(model::Type{<:TimeSeriesModel{D}}, timeseries, Δ; lowerΩcutoff, upperΩcutoff, taper)

Create storage for timeseries data in a format which is useful for Whittle methods.

Base.:+ — Method

M₁::Type{<:TimeSeriesModel{D,T}} + M₂::Type{<:TimeSeriesModel{D,T}} -> AdditiveTimeSeriesModel{M₁,M₂,D,T}

Base.:+ — Method

M₁::Type{<:TimeSeriesModel{D,T}} + M₂::Type{<:TimeSeriesModel{D,T}} -> Type{AdditiveTimeSeriesModel{M₁,M₂,D,T}}

Base.ndims — Method

ndims(::TimeSeriesModel) -> Integer
ndims(::Type{<:TimeSeriesModel}) -> Integer

Return the dimension of a timeseries model.

WhittleLikelihoodInference.EI! — Method

EI!(store, model::TimeSeriesModel)

Compute the expected periodogram and assign to appropriate place in memory (note this computes at twice the desired resolution).

WhittleLikelihoodInference.EI — Method

EI(model::TimeSeriesModel, n, Δ)

Compute EI at Fourier frequencies fftshift(fftfreq(n,2π/Δ)).

Note internal computation provides values at twice the resolution, this function returns at the desired resolution.

WhittleLikelihoodInference.acv! — Method

acv!(out, model::TimeSeriesModel, τ)

Compute the acv for a single lag and when acv is known.

WhittleLikelihoodInference.acv! — Method

acv!(store::Acv2EIStorage, model::TimeSeriesModel, encodedtime::LagsEI)

Compute the acv at many lags and allocate to storage when acv is known.

WhittleLikelihoodInference.acv! — Method

acv!(store::AdditiveStorage, model::AdditiveTimeSeriesModel)

Compute the acv for an additive model.

Processes storage and model, and computes acv for each separately, then combines and stores in the leftmost storage. In other words, the sum of both autocovariances is stores in store 1. This is preferable as for some models the acv may be known.

WhittleLikelihoodInference.acv! — Method

acv!(store::Sdf2EIStorage, model::UnknownAcvTimeSeriesModel, encodedtime::FreqAcvEst)

Approximate the acv and allocate to storage when acv is unknown.

WhittleLikelihoodInference.acv! — Method

acv!(store::TimeSeriesModelStorageFunction, model::TimeSeriesModel)

Unwrap storage and pass to lower level acv! call.

WhittleLikelihoodInference.acv! — Method

acv!(store::Acv2EIStorageUni, model::TimeSeriesModel{1,T}, encodedtime::LagsEI)

Compute the acv at many lags and allocate to storage when acv is known in the univariate case.

WhittleLikelihoodInference.acv! — Method

acv!(store::Sdf2EIStorageUni, model::UnknownAcvTimeSeriesModel{1,T}, encodedtime::FreqAcvEst)

Approximate the acv and allocate to storage when acv is known in the univariate case.

WhittleLikelihoodInference.acv — Method

acv(model::TimeSeriesModel, n, Δ)

Compute acv at lags -(n-1)*Δ:Δ:(n-1)*Δ.

WhittleLikelihoodInference.acv — Method

acv(model::TimeSeriesModel, τ)

Compute acv at τ, provided acv is known.

WhittleLikelihoodInference.add_asdf! — Method

add_asdf!(out, model::TimeSeriesModel, ω, Δ)

Add the aliased sdf of a model to a storage vector.

WhittleLikelihoodInference.add_sdf! — Method

add_sdf!(out, model::TimeSeriesModel, ω)

Add the sdf of a model to a storage vector.

WhittleLikelihoodInference.asdf! — Method

asdf!(out, model::TimeSeriesModel, ω, Δ)

Computes the asdf and overwrites a storage vector.

WhittleLikelihoodInference.asdf! — Method

asdf!(store::TimeSeriesModelStorage, model::TimeSeriesModel, freq::FreqAcvEst)

Compute the asdf for all frequencies and allocates to appropriate location in storage.

WhittleLikelihoodInference.asdf — Method

asdf(model::TimeSeriesModel, ω, Δ)

Compute asdf with sampling rate Δ at ω.

WhittleLikelihoodInference.checkparameterlength — Method

checkparameterlength(x,model::Type{<:TimeSeriesModel})

checks if the parameter vector is the correct length for the given model.

WhittleLikelihoodInference.checkposreal — Method

checkposreal(x)

Checks if x is a positive real number.

WhittleLikelihoodInference.coherance — Method

coherance(model::TimeSeriesModel, ω)

Compute the coherance of a given model at frequency ω.

WhittleLikelihoodInference.coherancy — Method

coherancy(model::TimeSeriesModel, ω)

Compute the coherancy of a given model at frequency ω.

WhittleLikelihoodInference.debiasedwhittle! — Method

debiasedwhittle!(store::TimeSeriesModelStorage, model::TimeSeriesModel, data::GenWhittleData)

Function to compute the debiased Whittle likelihood using a preallocated store.

WhittleLikelihoodInference.debiasedwhittle_Ehess! — Method

debiasedwhittle_Ehess!(EH, store, data::GenWhittleData)

Compute the expected hessian of the de-biased Whittle likelihood.

WhittleLikelihoodInference.debiasedwhittle_FG! — Method

debiasedwhittle_FG!(F, G, store, model::TimeSeriesModel, data::GenWhittleData)

Compute the debiased Whittle likelihood and its gradient using a preallocated store.

WhittleLikelihoodInference.debiasedwhittle_FGH! — Method

debiasedwhittle_FGH!(F, G, H, store, model::TimeSeriesModel, data::GenWhittleData)

Compute the debiased Whittle likelihood and its gradient and hessian using a preallocated store.

WhittleLikelihoodInference.debiasedwhittle_Fisher! — Method

debiasedwhittle_Fisher!(F, G, H, store, model::TimeSeriesModel, data::GenWhittleData)

Compute the debiased Whittle likelihood, its gradient and fisher information.

WhittleLikelihoodInference.extract_EI — Method

extract_EI(store::Sdf2EIStorage)
extract_EI(store::Acv2EIStorage)
extract_EI(store::Sdf2EIStorageUni)
extract_EI(store::Acv2EIStorageUni)

Extract the EI from general storage.

WhittleLikelihoodInference.extract_acv — Method

extract_acv(store::Sdf2EIStorage)
extract_acv(store::Acv2EIStorage)
extract_acv(store::Sdf2EIStorageUni)
extract_acv(store::Acv2EIStorageUni)

Extract the acv from general storage.

WhittleLikelihoodInference.extract_array — Method

extract_array(store::Sdf2EIStorage)
extract_array(store::Acv2EIStorage)
extract_array(store::Sdf2EIStorageUni)
extract_array(store::Acv2EIStorageUni)

Extract array from storage.

WhittleLikelihoodInference.fit — Method

fit(ts,Δ;model::Type{<:TimeSeriesModel},x₀,lowerΩcutoff,upperΩcutoff,x_lowerbounds,x_upperbounds,method,taper)
fit(timeseries::TimeSeries;model::Type{<:TimeSeriesModel},x₀,lowerΩcutoff,upperΩcutoff,x_lowerbounds,x_upperbounds,method,taper)

Fit a time series model using the IPNewton method from Optim.jl.

Arguments

ts: n by D matrix containing the timeseries (or vector if D=1), where n is the number of observations and D is the number of series.

Δ: The sampling rate, which should be a positive real number.

timeseries: Can be provided in place of ts and Δ.

model: The model which will be fitted. Should be a type (not a realisation of the model) e.g. JONSWAP{k} not JONSWAP{K}(x).

x₀: The initial parameter guess.

lowerΩcutoff: The lower cutoff for the frequency range to be used in fitting. Default is 0.

upperΩcutoff: The upper cutoff for the frequency range to be used in fitting. Default is Inf.

x_lowerbounds: The lower bounds on the parameter space. If nothing is provided (the default) then these are set to default values based on the model.

x_upperbounds: The upper bounds on the parameter space. If nothing is provided (the default) then these are set to default values based on the model.

method: Either :Whittle or :debiasedWhittle.

taper: The choice of tapering to be used. This should be nothing (in which case no taper is used) or dpss_nw where nw time-bandwith product (see DSP.dpss for more details).

options: Options for optimisation of type Optim.Options.

WhittleLikelihoodInference.getallderiv — Method

getallderiv(store::AdditiveStorage)
getallderiv(store::TimeSeriesModelStorageGradient)

Extract all of the derivatives from a generic store.

WhittleLikelihoodInference.grad_EI! — Method

grad_EI!(store::TimeSeriesModelStorageGradient, model::TimeSeriesModel)

Compute the gradient of the expected periodogram and assign to appropriate place in memory (note this computes at twice the desired resolution).

WhittleLikelihoodInference.grad_acv! — Method

grad_acv!

Compute the gradient of the acv and allocate to storage as appropriate.

WhittleLikelihoodInference.grad_acv — Method

grad_acv(model::TimeSeriesModel, τ)

Compute the gradient of the acv at τ.

WhittleLikelihoodInference.grad_add_asdf! — Method

grad_add_asdf!(out, model::TimeSeriesModel, ω, Δ)

Add the gradient of the aliased sdf of a model to an input storage vector.

WhittleLikelihoodInference.grad_add_sdf! — Method

grad_add_sdf!(out, model::TimeSeriesModel, ω)

Add the gradient of the sdf of a model to an input storage vector.

WhittleLikelihoodInference.grad_asdf! — Method

grad_asdf!(out, model::TimeSeriesModel, ω, Δ)

Compute the gradient of the aliased sdf of a model and store in storage vector.

WhittleLikelihoodInference.grad_asdf! — Method

grad_asdf!(store::TimeSeriesModelStorage, model::TimeSeriesModel, freq::FreqAcvEst)
grad_asdf!(store::TimeSeriesModelStorageGradient, model::TimeSeriesModel)

Compute the gradient of the asdf for all frequencies and allocate to appropriate location in storage.

WhittleLikelihoodInference.grad_sdf — Method

grad_sdf(model::TimeSeriesModel, ω)

Compute the gradient of the sdf at ω.

WhittleLikelihoodInference.groupdelay — Method

groupdelay(model::TimeSeriesModel, ω)

Compute the groupdelay of a given model at frequency ω.

WhittleLikelihoodInference.hess_EI! — Method

hess_EI!(store::TimeSeriesModelStorageHessian, model::TimeSeriesModel)

Compute the Hessian of the expected periodogram and assign to appropriate place in memory (note this computes at twice the desired resolution).

WhittleLikelihoodInference.hess_acv! — Method

hess_acv!

Compute the Hessian of the acv and allocate to storage as appropriate.

WhittleLikelihoodInference.hess_add_asdf! — Method

hess_add_asdf!(out, model::Type{<:TimeSeriesModel}, ω, Δ)

Add the Hessian of the aliased sdf of a model to an input storage vector.

WhittleLikelihoodInference.hess_add_sdf! — Method

hess_add_sdf!(out, model::TimeSeriesModel, ω)

Add the Hessian of the sdf of a model to an input storage vector.

WhittleLikelihoodInference.hess_asdf! — Method

hess_asdf!(out, model::TimeSeriesModel, ω, Δ)

Compute the Hessian of the aliased sdf of a model and store in storage vector.

WhittleLikelihoodInference.hess_asdf! — Method

hess_asdf!(store::TimeSeriesModelStorage, model::TimeSeriesModel, freq::FreqAcvEst)
hess_asdf!(store::TimeSeriesModelStorageHessian, model::TimeSeriesModel)

Compute the Hessian of the asdf for all frequencies and allocate to appropriate location in storage.

WhittleLikelihoodInference.hess_sdf — Method

hess_sdf(model::TimeSeriesModel, ω)

Compute the Hessian of the sdf at ω.

WhittleLikelihoodInference.indexLT — Method

indexLT(i,j,d)

Return the lower triangle index for an elements of a symmetric matrix assuming the lower triangle is encoded down the columns.

WhittleLikelihoodInference.minbins — Method

minbins(model::TimeSeriesModel)

Returns the minimum number of bins required for a good approximation of the sdf from the acv for a given model.

WhittleLikelihoodInference.nalias — Method

nalias(model::TimeSeriesModel)

Returns the number of times a sdf should be alised for a given model.

WhittleLikelihoodInference.nlowertriangle_dimension — Method

nlowertriangle_dimension(::TimeSeriesModel) -> Integer
nlowertriangle_dimension(::Type{<:TimeSeriesModel}) -> Integer

Return the number of elements in the lower triangle the spectral density matrix function of a timeseries model.

WhittleLikelihoodInference.nlowertriangle_parameter — Method

nlowertriangle_parameter(::TimeSeriesModel) -> Integer
nlowertriangle_parameter(::Type{<:TimeSeriesModel}) -> Integer

Return the number of elements in the lower triangle of the hessian matrix for a model with given number of parameters.

WhittleLikelihoodInference.parameternames — Method

parameternames(::TimeSeriesModel)
parameternames(::Type{<:TimeSeriesModel})

Return the parameter names for a given timeseries model.

WhittleLikelihoodInference.plotacv — Function

plotacv(model)
plotacv!(model)
plotacv(model, τ)
plotacv!(model, τ)
plotacv(model, n, Δ)
plotacv!(model, n, Δ)

Plot the aliased spectral density function of a model at lags τ. If the model does not have known acv, then the number of lags n and sampling rate Δ should be provided. In this case, the lags are -(n-1)*Δ:Δ:(n-1)*Δ. If unprovided, n = 30 and Δ = 1.

WhittleLikelihoodInference.plotasdf — Function

plotasdf(model)
plotasdf!(model)
plotasdf(model, Ω, Δ)
plotasdf!(model, Ω, Δ)

Plot the aliased spectral density function of a model at the frequencies Ω. By default, Ω = range(-π,π,length=200) and Δ = 1.

WhittleLikelihoodInference.plotcoh — Function

plotcoh(model)
plotcoh!(model)
plotcoh(model, Ω)
plotcoh!(model, Ω)

Plot the coherance and group delay of a model at the frequencies Ω. If multivariate, above the diagonal is the coherance, and below is the group delay and on the diagonal is the spectral density function. By default, Ω = range(-π,π,length=200) and Δ = 1.

WhittleLikelihoodInference.plotei — Function

plotei(model)
plotei!(model)
plotei(model, n, Δ)
plotei!(model, n, Δ)

Plot the aliased spectral density function of a model at the angular Fourier frequencies fftfreq(n,2π/Δ). By default, n = 1024 and Δ = 1.

WhittleLikelihoodInference.plotsdf — Function

plotsdf(model)
plotsdf!(model)
plotsdf(model, Ω)
plotsdf!(model, Ω)

Plot the spectral density function of a model at the frequencies Ω. By default, Ω = range(-π,π,length=200) and Δ = 1.

WhittleLikelihoodInference.sdf — Method

sdf(model::TimeSeriesModel, ω)

Compute sdf at ω.

WhittleLikelihoodInference.whittle! — Method

whittle!(store, model::TimeSeriesModel, data::GenWhittleData)

Compute the Whittle likelihood using a preallocated store.

WhittleLikelihoodInference.whittle_FG! — Method

whittle_FG!(F, G, store, model::TimeSeriesModel, data::GenWhittleData)

Compute the Whittle likelihood and its gradient using a preallocated store.

WhittleLikelihoodInference.whittle_FGH! — Method

whittle_FGH!(F, G, H, store, model::TimeSeriesModel, data::GenWhittleData)

Compute the debiased Whittle likelihood and its gradient and hessian using a preallocated store.