Fitting

The fit function can be used to fit a model to a recorded series with either the Whittle or debiased Whittle likelihoods using the interior point Newton method as implemented in Optim.jl.

First load the package:

using OceanWaveSpectralFitting

As a basic example, consider a univariate Gaussian process with JONSWAP spectral density function. We can simulate such a process with the following:

import Random # hide
Random.seed!(1234) # hide
using OceanWaveSpectralFitting # hide
n = 2304
Δ = 1/1.28
nreps = 1
α = 0.7
ωₚ = 1.1
γ = 3.3
r = 5.0
ts = simulate_gp(JONSWAP{1}(α,ωₚ,γ,r),n,Δ,nreps)[1]

In this case, we have simulated a Gaussian process with JONSWAP spectral density function with parameters α = 0.7, ωₚ = 1.1, γ = 3.3, r = 5.0 of length 2304 sampled every 1/1.28 seconds. This corresponds to 30 minutes of data. The function simulate_gp will simulate a vector of nreps series, which is why we recover the first of these to get one series.

We can now fit a model with the following:

import Random # hide
Random.seed!(1234) # hide
using OceanWaveSpectralFitting # hide
n = 2304 # hide
Δ = 1/1.28 # hide
nreps = 1 # hide
α = 0.7 # hide
ωₚ = 1.1 # hide
γ = 3.3 # hide
r = 5.0 # hide
ts = simulate_gp(JONSWAP{1}(α,ωₚ,γ,r),n,Δ,nreps)[1] # hide
x₀ = [α,ωₚ,γ,r] .+ 0.1
res = fit(ts,model=JONSWAP{1},x₀=x₀)

Here x₀ is the vector of initial parameter guesses. The estimated parameter vector can be recovered by doing:

import Random # hide
Random.seed!(1234) # hide
using OceanWaveSpectralFitting # hide
n = 2304 # hide
Δ = 1/1.28 # hide
nreps = 1 # hide
α = 0.7 # hide
ωₚ = 1.1 # hide
γ = 3.3 # hide
r = 5.0 # hide
ts = simulate_gp(JONSWAP{1}(α,ωₚ,γ,r),n,Δ,nreps)[1] # hide
x₀ = [α,ωₚ,γ,r] .+ 0.1 # hide
res = fit(ts,model=JONSWAP{1},x₀=x₀) # hide
x̂ = res.minimizer

The full example is:

import Random # hide
Random.seed!(1234) # hide
using OceanWaveSpectralFitting
n = 2304
Δ = 1/1.28
nreps = 1
α = 0.7
ωₚ = 1.1
γ = 3.3
r = 5.0
ts = simulate_gp(JONSWAP{1}(α,ωₚ,γ,r),n,Δ,nreps)[1]
x₀ = [α,ωₚ,γ,r] .+ 0.1
res = fit(ts,model=JONSWAP{1},x₀=x₀)
x̂ = res.minimizer

For more options, see below:

WhittleLikelihoodInference.fit — Function

fit(ts,Δ;model::Type{<:TimeSeriesModel},x₀,lowerΩcutoff,upperΩcutoff,x_lowerbounds,x_upperbounds,method,taper)
fit(timeseries::TimeSeries;model::Type{<:TimeSeriesModel},x₀,lowerΩcutoff,upperΩcutoff,x_lowerbounds,x_upperbounds,method,taper)

Fit a time series model using the IPNewton method from Optim.jl.

Arguments

ts: n by D matrix containing the timeseries (or vector if D=1), where n is the number of observations and D is the number of series.

Δ: The sampling rate, which should be a positive real number.

timeseries: Can be provided in place of ts and Δ.

model: The model which will be fitted. Should be a type (not a realisation of the model) e.g. JONSWAP{k} not JONSWAP{K}(x).

x₀: The initial parameter guess.

lowerΩcutoff: The lower cutoff for the frequency range to be used in fitting. Default is 0.

upperΩcutoff: The upper cutoff for the frequency range to be used in fitting. Default is Inf.

x_lowerbounds: The lower bounds on the parameter space. If nothing is provided (the default) then these are set to default values based on the model.

x_upperbounds: The upper bounds on the parameter space. If nothing is provided (the default) then these are set to default values based on the model.

method: Either :Whittle or :debiasedWhittle.

taper: The choice of tapering to be used. This should be nothing (in which case no taper is used) or dpss_nw where nw time-bandwith product (see DSP.dpss for more details).

options: Options for optimisation of type Optim.Options.